+++ devel/flx/flx.h
@@ -1375,6 +1375,7 @@ namespace flx {
 #  error
 # endif
 
+	// Unlike 'exec_iflow_friver___', this one does NOT send a 'EOF' after command was executed (at least under POSIX).
 	class exec_ioflow_driver___
 	: public _ioflow_driver___<>
 	{
@@ -2091,9 +2092,11 @@ namespace flx {
 
 	typedef exec_iflow_driver___ rExecRDriver;
 	typedef exec_oflow_driver___ rExecWDriver;
+	typedef exec_ioflow_driver___ rExecDriver;
 
 	typedef exec_iflow___ rExecRFlow;
 	typedef exec_oflow___ rExecWFlow;
+	typedef exec_ioflow_driver___ rExecFlow;
 
 	extern fdr::rRDriver &VoidRDriver;
 	extern fdr::rWDriver &VoidWDriver;