+++ tools/msfgq/GUI/main.h
@@ -48,37 +48,6 @@ namespace main {
   };
 
   extern sclx::action_handler<sSession> Core;
-
-
-  struct rXMelody {
-  public:
-    melody::wMelody Melody;
-    mscmld::sRow Row;
-    bso::sU8 BaseOctave;
-    mscmld::eAccidental Accidental; // Accidental to use for alterate notes when in C key.
-    bso::sBool Overwrite;
-    void reset(bso::sBool P = true)
-    {
-      Melody.reset(P);
-      Row = qNIL;
-      BaseOctave = 0;
-      Accidental = mscmld::a_Undefined;
-      Overwrite = false;
-    }
-    qCDTOR(rXMelody);
-    void Init(void)
-    {
-      Melody.Init();
-      Row = qNIL;
-      BaseOctave = 2;
-      Accidental = mscmld::aSharp;
-      Overwrite = false;
-    }
-  };
-
-  typedef mtx::rHandle hGuard;
-
-  rXMelody &Get(hGuard &Guard);
 }
 
 #endif // MAIN_INC_